+++ mmwis/compile.sh
@@ -19,7 +19,7 @@ fi
 # compile mmwis and hils
 
 rm -rf deploy
-# rm -rf build
+rm -rf build
 mkdir build
 cd build
 cmake ../ -DCMAKE_C_COMPILER=$(which gcc) -DCMAKE_CXX_COMPILER=$(which g++) -DCMAKE_BUILD_TYPE=${buildtype}
@@ -28,7 +28,9 @@ cd ..
 
 mkdir deploy
 cp ./build/mmwis                                                         deploy/mmwis
-cp ./build/extern/struction/branch_reduce_convergence                    deploy/struction
-# rm -rf build
+if [ -f ./build/extern/struction/branch_reduce_convergence ]; then
+        cp ./build/extern/struction/branch_reduce_convergence                    deploy/struction
+fi
+rm -rf build
